Added pause/resume functionality Now retrieves song name from
Added ToneAC_RTTTL.ino application that demonstrates the usage of the library. It uses my NESController library to allow you to control playback with an old NES controller. If you for some reason don't have access to one of those to wire in it will still loop through all the songs.